Unify project used for packaging Self-contained Azure Function binaries (which now is GeneXus.Deploy.AzureFunctions.Handlers.csproj) with the project used to build GeneXus.Deploy.AzureFunctions.Handlers.dll (which is GeneXus.Deploy.AzureFunctions.Handlers.csproj + References.targets). Package reference to Microsoft.Azure.Functions.Worker.Extensions.ApplicationInsights is twice (in GeneXus.Deploy.AzureFunctions.Handlers.csproj and References.targets) because the second one overrides the ExcludeAssets (to fix packaging error NU5104 with that preview package: A stable release of a package should not have a prerelease dependency)
