+++ dotnet/src/extensions/Azure/Handlers/Annotations/AnnotationsPlaceHolder.cs
@@ -2,7 +2,7 @@ using System.Net;
 using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.Functions.Worker.Http;
 
-namespace HttpTriggerDummy
+namespace AllFunctions
 {
 	// This is a dummy class.
 	// The purpose of this class is to force the Net SDK to generate all the configuration files
@@ -20,6 +20,30 @@ namespace HttpTriggerDummy
 			response.WriteString("Welcome to Azure Functions!");
 
 			return response;
+		}
+	}
+	public static class QueueTriggerDummy
+	{
+		[Function("QueueTrigger1")]
+		public static void Run([QueueTrigger("myqueue-items", Connection = "")] string myQueueItem,
+			FunctionContext context)
+		{
+
+		}
+	}
+	public static class ServiceBusQueueDummy
+	{
+		[Function("ServiceBusQueueTrigger1")]
+		public static void Run([ServiceBusTrigger("myqueue", Connection = "")] string myQueueItem, FunctionContext context)
+		{
+
+		}
+	}
+	public static class TimerTriggerDummy
+	{
+		[Function("TimerTrigger1")]
+		public static void Run([TimerTrigger("0 */5 * * * *")] string myTimer, FunctionContext context)
+		{
 		}
 	}
 }
