+++ dotnet/src/extensions/Azure/Handlers/DummyProgram.cs
@@ -1,0 +1,18 @@
+using System.Threading.Tasks;
+using Microsoft.Extensions.Configuration;
+using Microsoft.Extensions.Hosting;
+using Microsoft.Azure.Functions.Worker.Configuration;
+
+namespace AllFunctions
+{
+    public class DummyProgram
+    {
+        public static void Main()
+        {
+            var host = new HostBuilder()
+                .ConfigureFunctionsWorkerDefaults()
+                .Build();
+            host.Run();
+        }
+    }
+}
