+++ dotnet/src/extensions/Azure/Handlers/GeneXus.Deploy.AzureFunctions.Handlers.csproj
@@ -1,55 +1,40 @@
-<Project Sdk="Microsoft.NET.Sdk">
-  <PropertyGroup>
-    <TargetFrameworks>net5.0;net6.0</TargetFrameworks>
-    <OutputType>Exe</OutputType>
-  </PropertyGroup>
-  <PropertyGroup Condition="$(TargetFramework) == 'net6.0'">
-	  <AzureFunctionsVersion>v4</AzureFunctionsVersion>
-  </PropertyGroup>
-  <PropertyGroup Condition="$(TargetFramework) == 'net5.0'">
-	  <AzureFunctionsVersion>v3</AzureFunctionsVersion>
-  </PropertyGroup>
-  <ItemGroup>
-    <PackageReference Include="log4net" Version="2.0.11" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.13" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.ServiceBus" Version="4.2.1" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage" Version="4.0.4" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Timer" Version="4.0.1" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.3.0" OutputItemType="Analyzer" Condition="$(TargetFramework) == 'net6.0'" />
-	<PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.0.3" OutputItemType="Analyzer" Condition="$(TargetFramework) == 'net5.0'" />
-	<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.5.2" Condition="$(TargetFramework) == 'net6.0'" />
-	<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.1.0" Condition="$(TargetFramework) == 'net5.0'" />
-    <PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.ApplicationInsights" Version="1.0.0-preview4" Condition="$(TargetFramework) == 'net6.0'" >
-		<ExcludeAssets>all</ExcludeAssets>
-	</PackageReference>
-    <PackageReference Include="Microsoft.Data.SqlClient" Version="1.0.19269.1" />
-    <PackageReference Include="StackExchange.Redis" Version="2.0.593" />  
-  </ItemGroup>
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\dotnetcore\GxClasses.Web\GxClasses.Web.csproj" />
-    <ProjectReference Include="..\..\..\dotnetcore\GxClasses\GxClasses.csproj" />
-    <ProjectReference Include="..\..\..\dotnetcore\GxExcel\GxExcel.csproj" />
-    <ProjectReference Include="..\..\..\dotnetcore\GxMail\GxMail.csproj" />
-    <ProjectReference Include="..\..\..\dotnetcore\Providers\Cache\GxRedis\GxRedis.csproj" />
-  </ItemGroup>
-  <ItemGroup>
-    <None Update="host.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-    </None>
-    <None Update="local.settings.json">
-      <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
-      <CopyToPublishDirectory>Never</CopyToPublishDirectory>
-    </None>
-  </ItemGroup>
-  <ItemGroup>
-    <Content Include="$(OutputPath)net6.0\$(MSBuildProjectName).deps.json">
-     <Pack>true</Pack>
-     <PackagePath>./lib/net6.0/</PackagePath>
-    </Content>
-    <Content Include="$(OutputPath)net5.0\$(MSBuildProjectName).deps.json">
-     <Pack>true</Pack>
-     <PackagePath>./lib/net5.0/</PackagePath>
-    </Content>
+﻿<Project Sdk="Microsoft.NET.Sdk">
+	<PropertyGroup>
+		<TargetFrameworks>net5.0;net6.0</TargetFrameworks>
+		<OutputType>Exe</OutputType>
+	</PropertyGroup>
+	<PropertyGroup Condition="$(TargetFramework) == 'net6.0'">
+		<AzureFunctionsVersion>v4</AzureFunctionsVersion>
+	</PropertyGroup>
+	<PropertyGroup Condition="$(TargetFramework) == 'net5.0'">
+		<AzureFunctionsVersion>v3</AzureFunctionsVersion>
+	</PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="Microsoft.AspNetCore.Http.Abstractions" Version="2.2.0" />
+		<PackageReference Include="Newtonsoft.Json" Version="13.0.1" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Http" Version="3.0.13" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.ServiceBus" Version="4.2.1" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Storage" Version="4.0.4" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.Timer" Version="4.0.1" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.3.0" OutputItemType="Analyzer" Condition="$(TargetFramework) == 'net6.0'" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Sdk" Version="1.0.3" OutputItemType="Analyzer" Condition="$(TargetFramework) == 'net5.0'" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.5.2" Condition="$(TargetFramework) == 'net6.0'" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker" Version="1.1.0" Condition="$(TargetFramework) == 'net5.0'" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Extensions.ApplicationInsights" Version="1.0.0-preview4" Condition="$(TargetFramework) == 'net6.0'"/>
+		<PackageReference Include="StackExchange.Redis" Version="2.0.593" />
+	</ItemGroup>
+	<PropertyGroup>
+		<GXclassesRefs>References.targets</GXclassesRefs>
+	</PropertyGroup>
+	<Import Condition="Exists('$(GXclassesRefs)')" Project="$(GXclassesRefs)" />
 
-  </ItemGroup>
+	<ItemGroup>
+		<None Update="host.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+		</None>
+		<None Update="local.settings.json">
+			<CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
+			<CopyToPublishDirectory>Never</CopyToPublishDirectory>
+		</None>
+	</ItemGroup>
 </Project>
