+++ dotnet/src/extensions/Azure/Handlers/Program.cs
@@ -9,7 +9,6 @@ using GxClasses.Web;
 using GxClasses.Web.Middleware;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
-using StackExchange.Redis;
 
 namespace GeneXus.Deploy.AzureFunctions.Handlers
 {
