+++ dotnet/src/extensions/Azure/Handlers/host.json
@@ -1,11 +1,16 @@
 {
-    "version": "2.0",
-    "logging": {
-        "applicationInsights": {
-            "samplingSettings": {
-                "isEnabled": true,
-                "excludedTypes": "Request"
-            }
-        }
-    }
+	"version": "2.0",
+	"logging": {
+		"applicationInsights": {
+			"samplingSettings": {
+				"isEnabled": true,
+				"excludedTypes": "Request"
+			}
+		}
+	},
+	"extensions": {
+		"http": {
+			"routePrefix": "rest"
+		}
+	}
 }
